Rename some classes
