+++ NovaCore-Spigot/src/main/java/net/zeeraa/novacore/spigot/module/modules/scoreboard/text/ModifyableTextLine.java
@@ -1,21 +1,27 @@
 package net.zeeraa.novacore.spigot.module.modules.scoreboard.text;
 
-import java.util.function.Supplier;
-
 import net.kyori.adventure.text.Component;
 import net.megavex.scoreboardlibrary.api.sidebar.component.SidebarComponent.Builder;
 
-public class DynamicText extends ScoreboardEntry {
-	private Supplier<String> supplier;
+public class ModifyableTextLine extends ScoreboardLine {
+	private String text;
 
-	public DynamicText(Supplier<String> supplier) {
-		this.supplier = supplier;
+	public ModifyableTextLine(String text) {
+		this.text = text;
+	}
+
+	public String getText() {
+		return text;
+	}
+
+	public void setText(String text) {
+		this.text = text;
 	}
 
 	@Override
 	public Builder apply(Builder builder) {
 		builder.addDynamicLine(() -> {
-			return Component.text(supplier.get());
+			return Component.text(text);
 		});
 		return builder;
 	}
