+++ NovaCore-Spigot/src/main/java/net/zeeraa/novacore/spigot/module/modules/scoreboard/text/ScoreboardLine.java
@@ -2,7 +2,7 @@ package net.zeeraa.novacore.spigot.module.modules.scoreboard.text;
 
 import net.megavex.scoreboardlibrary.api.sidebar.component.SidebarComponent;
 
-public abstract class ScoreboardEntry {
+public abstract class ScoreboardLine {
 	public abstract SidebarComponent.Builder apply(SidebarComponent.Builder builder);
 	
 	public void tick() {
