+++ NovaCore-Spigot/src/main/java/net/zeeraa/novacore/spigot/module/modules/scoreboard/text/StaticTextLine.java
@@ -3,26 +3,20 @@ package net.zeeraa.novacore.spigot.module.modules.scoreboard.text;
 import net.kyori.adventure.text.Component;
 import net.megavex.scoreboardlibrary.api.sidebar.component.SidebarComponent.Builder;
 
-public class ModifyableText extends ScoreboardEntry {
+public class StaticTextLine extends ScoreboardLine {
 	private String text;
 
-	public ModifyableText(String text) {
+	public StaticTextLine(String text) {
 		this.text = text;
 	}
 
 	public String getText() {
 		return text;
 	}
-
-	public void setText(String text) {
-		this.text = text;
-	}
-
+	
 	@Override
 	public Builder apply(Builder builder) {
-		builder.addDynamicLine(() -> {
-			return Component.text(text);
-		});
+		builder.addStaticLine(Component.text(text));
 		return builder;
 	}
 }
